The final documentation for the Data warehouse project
This group project was made for the Data warehouse and Business Intelligence class during my Bachelor's Degree. The project was done using Ms SQL Server, Visual Studio and Power BI. The general purpose of the project was to create a Data warehouse to store and update data from the chosen data sources, which was then used to generate reports.