Code for dashboard
This group project was made for the Exploration and analysis of data class during my Bachelor's Degree studies. The main aim of the project was to create an interactive dashboard using Shiny package in R based on the chosen dataset. Our project focuses on the carbon footprint of different models of cars registered in Poland during the period 2015-2021. Each member of the group made a single page with a graph, the author of the second one is me.